requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
selenium==4.12.0
webdriver-manager==4.0.0
lxml==4.9.3 